+++ src/app/components/content-spaces-new.js
@@ -7,7 +7,7 @@
  */
 
 import Component from '@ember/component';
-import I18n from 'onedata-gui-common/mixins/components/i18n';
+import I18n from 'onedata-gui-common/mixins/i18n';
 import { inject } from '@ember/service';
 import globals from 'onedata-gui-common/utils/globals';