+++ src/app/components/harvester-remove-modal.js
@@ -8,7 +8,7 @@
 
 import Component from '@ember/component';
 import { observer } from '@ember/object';
-import I18n from 'onedata-gui-common/mixins/components/i18n';
+import I18n from 'onedata-gui-common/mixins/i18n';
 import notImplementedThrow from 'onedata-gui-common/utils/not-implemented-throw';
 
 export default Component.extend(I18n, {