+++ src/app/components/modals/api-samples-modal.js
@@ -9,7 +9,7 @@
 import Component from '@ember/component';
 import { inject as service } from '@ember/service';
 import { computed } from '@ember/object';
-import I18n from 'onedata-gui-common/mixins/components/i18n';
+import I18n from 'onedata-gui-common/mixins/i18n';
 import { reads } from '@ember/object/computed';
 import { promise } from 'ember-awesome-macros';
 import recordIcon from 'onedata-gui-common/utils/record-icon';