+++ src/app/components/provider-info-name.js
@@ -6,7 +6,7 @@
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 import Component from '@ember/component';
-import I18n from 'onedata-gui-common/mixins/components/i18n';
+import I18n from 'onedata-gui-common/mixins/i18n';
 export default Component.extend(I18n, {
   tagName: 'span',
   classNames: ['provider-info-name'],