+++ src/app/locales/en/tabs/atm-inventories.js
@@ -1,5 +1,9 @@
 export default {
   menuItem: 'Automation',
+  specialResources: {
+    empty: 'No automation inventories',
+    new: 'Create new automation inventory',
+  },
   aspects: {
     workflows: 'Workflows',
     lambdas: 'Lambdas',