+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-10262 Added content names for special resources in mobile mode
 * VFS-11810 Changed global file URL to be prettier
 * VFS-11685 Added loading and changing effective privileges during modification
 * VFS-11684 Added tooltips for effective privileges